requirements.txt
tensorflow-gpu
opencv-python
matplotlib
gym==0.9.1
mujoco-py==0.5.7
scipy
